Changes were made to make the uCOS build
Changes were made to the port files and all the os files were placed in a single folder. The test.mcp project was imported and the os is able to build now. All uneccessary files in the port folders were removed.

An app_cfg.h file was created to define the stack size and other things.